+++ src/ruby/models/circle.rb
@@ -50,4 +50,46 @@ class Circle
       gradient: gradient&.to_h
     }
   end
+
+  # Returns `true` if the {Vector2} is contained inside this {Circle} or is
+  # on the border of.
+  #
+  # @example Basic usage
+  #   hitbox = Circle.new(x: 10, y: 10, radius: 10)
+  #   position = Vector2.new(x: 0, y: 0)
+  #
+  #   puts hitbox.overlaps?(position)
+  #   # => false
+  #
+  #   # The position changes to inside the circle...
+  #   position = Vector2.new(x: 15, y: 15)
+  #
+  #   puts hitbox.overlaps?(position)
+  #   # => true
+  #
+  #   # The position changes to the edge of the circle...
+  #   position = Vector2.new(x: 0, y: 10)
+  #
+  #   puts hitbox.overlaps?(position)
+  #   # => true
+  #
+  #   # The position changes to far outside circle...
+  #   position = Vector2.new(x: 100, y: 100)
+  #
+  #   puts hitbox.overlaps?(position)
+  #   # => false
+  #
+  # @param other [Vector2]
+  # @return [Boolean]
+  # @raise [ArgumentError] If passed anything other than a Vector2
+  def overlaps?(other)
+    raise ArgumentError, "Must pass in a Vector2" unless other.is_a?(Vector2)
+
+    x_distance = x - other.x
+    y_distance = y - other.y
+
+    distance = Math.sqrt((x_distance**2) + (y_distance**2))
+
+    distance <= radius
+  end
 end